Fix security and JWT token handling, update CORS configuration, and improve user creation flow
- Updated CORS configuration to allow frontend requests.
- Adjusted security settings, including JWT token handling.
- Improved user creation flow for SelfTrainer.
- Added secure password storage and authentication validation in the login process.

+++ src/main/java/com/trainWise/app/model/TrainingUnit.java
@@ -29,9 +29,9 @@ public class TrainingUnit {
 
     private int set;
 
-    private double weight;
+    private int weight;
 
-    public TrainingUnit( Exercise exercise, int reps, int set, double weight) {
+    public TrainingUnit( Exercise exercise, int reps, int set, int weight) {
         this.exercise = exercise;
         this.reps = reps;
         this.set = set;
